Improved generating of JSONs
JSONs will now have "_id" instead of "Id" for better importing data into db
