+++ backendDotnet/DatabaseSerializer/Program.cs
@@ -1,5 +1,4 @@
 ﻿using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using System.Reflection;
 
 namespace DatabaseSerializer
@@ -16,10 +15,10 @@ namespace DatabaseSerializer
             var modelTypesArray = modelTypes.ToArray();
 
             var outputDirModelTypes = Path.Combine(Assembly.GetExecutingAssembly().Location, "..", "..", "..", "ModelsTypes");
-            SaveModels(outputDirModelTypes, modelTypesArray, assembly, new KeysJsonConverter(modelTypesArray));
+            SaveModels(outputDirModelTypes, modelTypesArray, assembly, new ModelTypesJsonConverter(modelTypesArray));
 
             var outputDirModels = Path.Combine(Path.GetDirectoryName(outputDirModelTypes), "ModelsExample");
-            SaveModels(outputDirModels, modelTypesArray, assembly, null);
+            SaveModels(outputDirModels, modelTypesArray, assembly, new ModelExamplesJsonConverter(modelTypesArray));
 
             Console.WriteLine("Models saved successfully.");
             Console.WriteLine($"They can be found in {Path.GetFullPath(Path.GetDirectoryName(outputDirModelTypes))}.");
@@ -56,71 +55,6 @@ namespace DatabaseSerializer
                     File.WriteAllText($"{path}/{t.Name}.json", jsonString);
                 }
             });
-        }
-    }
-
-    public class KeysJsonConverter : JsonConverter
-    {
-        private readonly Type[] _types;
-
-        public KeysJsonConverter(params Type[] types)
-        {
-            _types = types;
-        }
-
-        public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer)
-        {
-            JTokenWriter jtokenWriter = new JTokenWriter();
-            jtokenWriter.WriteStartObject();
-
-            foreach (var prop in value.GetType().GetProperties())
-            {
-                jtokenWriter.WritePropertyName(prop.Name);
-
-                var propertyType = prop.PropertyType;
-                var propertyTypeName = propertyType.FullName;
-                if (propertyType.IsGenericType)
-                {
-                    propertyTypeName = propertyType.Name + $"[{string.Join<string>(',', propertyType.GenericTypeArguments.Select(a=> a.FullName))}]";
-                }
-
-                if (propertyType.IsArray)
-                {
-                    jtokenWriter.WriteValue($"{propertyTypeName}");
-                }
-                else
-                {
-                    jtokenWriter.WriteValue(propertyTypeName);
-                }
-            }
-
-            jtokenWriter.WriteEndObject();
-
-            //jtokenWriter.WriteWhitespace(Environment.NewLine);
-            //JToken t = JToken.FromObject(value);
-
-
-            JObject o = jtokenWriter.Token as JObject;
-            if (o != null)
-            {
-                o.WriteTo(writer);
-                //t.WriteTo(writer);
-            }
-        }
-
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-        {
-            throw new NotImplementedException("Unnecessary because CanRead is false. The type will skip the converter.");
-        }
-
-        public override bool CanRead
-        {
-            get { return false; }
-        }
-
-        public override bool CanConvert(Type objectType)
-        {
-            return _types.Any(t => t == objectType);
         }
     }
 }
